+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Display an error message if the parameter `database` is not defined
 - Move the compiler binary to a Symfony command
 - Compatibility with Symfony 6 components
+- Rename SqlDumper class to MysqlDumper to avoid confusion
 
 ## [3.0.0] - 2022-07-28
 [3.0.0]: https://github.com/Smile-SA/gdpr-dump/compare/2.3.3...3.0.0